Tidy `overShelleyBasedLedgerState`.
Most of the functionality created in `Shelley/ShelleyBased` is only used
once, in `Cardano/ShelleyBased`, and is in fact only used for one
function. So localise everything to that module.
