+++ ouroboros-consensus-shelley/ouroboros-consensus-shelley.cabal
@@ -50,7 +50,6 @@ library
                        Ouroboros.Consensus.Shelley.Protocol.Abstract
                        Ouroboros.Consensus.Shelley.Protocol.Praos
                        Ouroboros.Consensus.Shelley.Protocol.TPraos
-                       Ouroboros.Consensus.Shelley.ShelleyBased
                        Ouroboros.Consensus.Shelley.ShelleyHFC
 
   build-depends:       base              >=4.9   && <4.15
